+++ src/Service/Lambda/src/Result/FunctionConfiguration.php
@@ -759,6 +759,7 @@ class FunctionConfiguration extends Result
     private function populateResultDurableConfig(array $json): DurableConfig
     {
         return new DurableConfig([
+            'KMSKeyArn' => isset($json['KMSKeyArn']) ? (string) $json['KMSKeyArn'] : null,
             'RetentionPeriodInDays' => isset($json['RetentionPeriodInDays']) ? (int) $json['RetentionPeriodInDays'] : null,
             'ExecutionTimeout' => isset($json['ExecutionTimeout']) ? (int) $json['ExecutionTimeout'] : null,
         ]);