+++ src/Service/Lambda/src/Result/ListVersionsByFunctionResponse.php
@@ -166,6 +166,7 @@ class ListVersionsByFunctionResponse extends Result implements \IteratorAggregat
     private function populateResultDurableConfig(array $json): DurableConfig
     {
         return new DurableConfig([
+            'KMSKeyArn' => isset($json['KMSKeyArn']) ? (string) $json['KMSKeyArn'] : null,
             'RetentionPeriodInDays' => isset($json['RetentionPeriodInDays']) ? (int) $json['RetentionPeriodInDays'] : null,
             'ExecutionTimeout' => isset($json['ExecutionTimeout']) ? (int) $json['ExecutionTimeout'] : null,
         ]);