+++ src/Service/Ec2/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: This launch surfaces the public SSM parameter associated with public AMIs in the AMI metadata.
+
 ## 1.1.0
 
 ### Added